Fixed DefaultRank/DefaultBuildRank/PatrolledRank not saving properly by ConfigGUI. Documented some player-related events better. Fixed the list of brush aliases not getting appended to brush help.

+++ fCraft/Drawing/DrawOperation.cs
@@ -28,7 +28,7 @@ namespace fCraft.Drawing {
                    BlocksSkipped,
                    BlocksTotalEstimate;
 
-        public float PercentDone {
+        public int PercentDone {
             get {
                 return (BlocksProcessed * 100) / BlocksTotalEstimate;
             }
